fix(dal): fix conversion problem from decimal to double
Decimal columns in database must map to decimal properties on models and projections.

+++ BLL/Projection/BookStoreProjection.cs
@@ -6,7 +6,7 @@ namespace BLL.Projection
   {
     public string Name { get; set; }
     public string OIB { get; set; }
-    public double DelayPricePerDay { get; set; }
+    public decimal DelayPricePerDay { get; set; }
     public string Email { get; set; }
   }
 }
